Workaround sphinx-click bug, split manage.py in two
sphinx-click has a bug due to it errors while trying to generate docs
for multicommand groups. Therefore, move actual commands to
listenbrainz.manage and add dump and spark commands in root manage.py.
